requirements.txt
Python == 3.5.2
numpy == 1.15.4
pandas == 0.23.4
torch == 1.0.1.post2
torchnet == 0.0.4
torchvision == 0.2.2.post3
scikit-learn == 0.19.1
scipy == 1.3.0
faiss-gpu == 1.5.3
numpy == 1.16.4
scikit-image == 0.15.0
Pillow == 6.0.0